refactor: move code from ModuleHelpers to GeneralModule that's only used for new user registration
