+++ MiraBot.Common/ModuleHelpers.cs
@@ -59,58 +59,9 @@ namespace MiraBot.Common
             }
         }
 
-        public async Task SaveUserTimezoneAsync(User owner)
+        public async Task ModifyUserAsync(User user)
         {
-            await SendTimezoneFileAsync();
-
-            while (true)
-            {
-                await ReplyAsync("Copy your timezone exactly how it's listed in this file, and send it to me so I can register your timezone!");
-
-                var input = await _interactive.NextMessageAsync(
-                    x => x.Author.Id == Context.User.Id && x.Channel.Id == Context.Channel.Id,
-                    timeout: TimeSpan.FromMinutes(2));
-
-                if (!input.IsSuccess)
-                {
-                    continue;
-                }
-
-                var timezone = input.Value.Content.Trim();
-
-                if (!IsValidTimezone(timezone))
-                {
-                    await ReplyAsync("The timezone you entered isn't valid! Make sure to copy your timezone exactly as it's listed in the file I sent.");
-                    continue;
-                }
-
-                if (owner.UsesAmericanDateFormat is null)
-                {
-                    await ReplyAsync("How would you write out the date \"August 30th\"?");
-                    var options = new List<string> { "8/30", "30/8" };
-                    await GenerateSelectMenuAsync(options,
-                        "How would you write out the date \"August 30th\"?",
-                        "select-menu",
-                        "Select this option",
-                        Context
-                        );
-                    var selection = result;
-                    result = -1;
-                    bool isAmerican = (selection == 0);
-
-                    await AddDateFormatToUserAsync(owner.DiscordId, isAmerican);
-                }
-
-                await AddTimezoneToUserAsync(Context.User.Id, timezone);
-                break;
-            }
-        }
-
-        public static bool IsValidTimezone(string timezoneId)
-        {
-            return TimeZoneInfo
-                .GetSystemTimeZones()
-                .Any(t => t.Id.Equals(timezoneId, StringComparison.OrdinalIgnoreCase));
+            await _usersRepository.ModifyUserAsync(user);
         }
 
 
@@ -218,51 +169,7 @@ namespace MiraBot.Common
                 await context.Channel.SendMessageAsync($"{counter}. **\"{message}\"**");
             }
         }
-
-        public async Task SendTimezoneFileAsync()
-        {
-            string fileName;
-            try
-            {
-                fileName = CreateTimezoneFile();
-            }
-            catch (Exception ex)
-            {
-                await ReplyAsync("I was unable to create the timezone file. Please seek assistance from the developer.");
-                return;
-            }
-
-            await FollowupWithFileAsync(fileName);
-        }
-
-        public async Task AddTimezoneToUserAsync(ulong discordId, string timezoneId)
-        {
-            var user = await _usersRepository.GetUserByDiscordIdAsync(discordId)
-                .ConfigureAwait(false);
-
-            if (user is not null)
-            {
-                user.Timezone = timezoneId;
-                await _usersRepository.ModifyUserAsync(user);
-            }
-        }
-
-        public async Task AddDateFormatToUserAsync(ulong discordId, bool isAmerican)
-        {
-            var user = await _usersRepository.GetUserByDiscordIdAsync(discordId)
-                .ConfigureAwait(false);
-
-            user.UsesAmericanDateFormat = isAmerican;
-            await _usersRepository.ModifyUserAsync(user);
-        }
-
-        public static string CreateTimezoneFile()
-        {
-            var fileName = Path.ChangeExtension(Path.GetRandomFileName(), ".txt");
-            var timeZones = TimeZoneInfo.GetSystemTimeZones();
-            File.WriteAllLines(fileName, timeZones.Select(t => t.Id).ToArray());
-            return fileName;
-        }
+ 
 
         public async Task<bool> UserExistsAsync(ulong discordId)
         {
